requirements.txt
playwright==1.47.0
tqdm==4.66.5
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.51.2
python-dotenv==1.0.1
pydantic==2.9.2
orjson==3.10.7
streamlit==1.39.0
